refactor: create pages directory and components directory
